bugs fixed

+++ backend/model/noteModel.js
@@ -5,8 +5,52 @@ const { ObjectId } = require("mongodb");
 class noteModel {
   async getAllNotes() {
     const db = await database.connectToServer();
-    const notes = db.collection("Notes").find({});
-    return await notes.toArray();
+
+    // Consulta para obtener todas las notas del usuario y sus subnotas
+    const notes = await db
+      .collection("Notes")
+      .find({ })
+      .toArray();
+
+    // Mapear notas a objetos con la estructura de árbol
+    const notesMap = {};
+    const rootNotes = [];
+
+    // Crear un mapa de notas por su ID y encontrar las notas raíz
+    notes.forEach((note) => {
+      notesMap[note._id] = note;
+
+      if (
+        note.referencedNoteId.equals(new ObjectId("000000000000000000000000"))
+      ) {
+        console.log("Funciona");
+        rootNotes.push(note);
+      }
+    });
+
+    // Función recursiva para obtener todas las subnotas de una nota dada
+    const getSubNotes = (noteId) => {
+      const subNotes = [];
+      notes.forEach((note) => {
+        if (note.referencedNoteId.equals(noteId)) {
+          const subNote = { ...note };
+          const childSubNotes = getSubNotes(note._id); // Llamada recursiva para obtener las subnotas de esta subnota
+          if (childSubNotes.length > 0) {
+            subNote.subNotes = childSubNotes; // Asignar las subnotas encontradas como subnotas de esta subnota
+          }
+          subNotes.push(subNote);
+        }
+      });
+
+      return subNotes;
+    };
+
+    // Asignar subnotas a las notas correspondientes
+    rootNotes.forEach((rootNote) => {
+      rootNote.subNotes = getSubNotes(rootNote._id);
+    });
+
+    return rootNotes;
   }
 
   async getNoteById(noteId, userId) {
@@ -55,13 +99,61 @@ class noteModel {
     return updateResult;
   }
 
-  async deleteNoteById(noteId) {
+  async deleteNoteById(noteId, userId, roleId) {
     const db = await database.connectToServer();
-    return await db
-      .collection("Notes")
-      .deleteOne({ _id: new ObjectId(noteId) });
-  }
+    const note = await db.collection("Notes").findOne({ _id: new ObjectId(noteId) })
+    if (note){
+      if (note.user_id === userId || roleId === 'a') {
+        return await db
+          .collection("Notes")
+          .deleteOne({ _id: new ObjectId(noteId) });
 
+      }
+    }else{
+      throw new Error("User does not have access to this note");
+    }
+  }
+  // async deleteNoteById(noteId, userId) {
+  //   const db = await database.connectToServer();
+  //   const notes = await this.getUserNotes(userId, noteId);
+  //   if (notes.length === 0) {
+  //     throw new Error("User notes not found");
+  //   }
+  
+  //   // Función recursiva para eliminar una nota y todas sus subnotas
+  //   const deleteNoteAndSubnotes = async (note) => {
+  //     // Si la nota tiene subnotas, eliminarlas recursivamente
+  //     if (note.subNotes && note.subNotes.length > 0) {
+  //       for (const subNote of note.subNotes) {
+  //         await deleteNoteAndSubnotes(subNote);
+  //       }
+  //     }
+  
+  //     // Eliminar la nota de la base de datos
+  //     await db.collection("Notes").deleteOne({ _id: note._id });
+  //   };
+  
+  //   // Función para encontrar la nota correspondiente al noteId en las notas y subnotas
+  //   const findNoteAndDelete = async (notes) => {
+  //     for (const note of notes) {
+  //       if (note._id === noteId) {
+  //         console.log(note)
+  //         if(note.user_id !== userId){
+  //           throw new Error("User does not have access to this note");
+  //         }
+  //         await deleteNoteAndSubnotes(note);
+  //         return; // Termina la búsqueda si la nota se encuentra y elimina
+  //       }
+  //       if (note.subNotes && note.subNotes.length > 0) {
+  //         await findNoteAndDelete(note.subNotes); // Busca en las subnotas recursivamente
+  //       }
+  //     }
+  //   };
+  
+  //   // Buscar y eliminar la nota correspondiente al noteId
+  //   await findNoteAndDelete(notes);
+  // }
+  
   // async getUserNotes(userId, referencedNoteId) {
 
   //     const db = await database.connectToServer();
